backend/requirements.txt
flask==1.1.2
PyYAML==5.1
psutil==5.7.3
